rtld-cheri-elf: let the kernel load relocbase into $c4
The kernel has that information easily available whereas it is quite hard
to do this in the assembly in a way that LLD will accept.
LLD refuses insert an absolute reloction into PIC binaries which is the
approach currently used to calculate the image base address. We could
modify lld to provide a symbol that can be loaded using PIC_LOAD_CODE_PTR
but as we already have the information available that seems overly
complicated.