HACK to work around a TLB exception when booting
When building the kernel with dev branch clang I get the following when
booting:
```
Trap cause = 2 (TLB miss (load or instr. fetch) - kernel mode)
[ thread pid 2 tid 100015 ]
Stopped at      ata_zac_mgmt_in+0x29c4: lbu     at,1016(v1)
```

The actual location is `path->device->inq_flags |= SID_DMA` inside
probefunc in ata_xpt.c. The crash doesn't happen with the nosp branch
but the only difference I can see in the generated code is that the dev
branch hoists the load of path->device before the if statement.
What's even weirder is that when I single step with GDB it will fail
when loading exactly the address that is printed by the printf().