Make truss(8) work for i686-unknown-cloudabi binaries on FreeBSD/amd64.
This change copies the existing amd64_cloudabi64.c to amd64_cloudabi32.c
and reimplements the functions for fetching system call arguments and
return values to use the same scheme as used by the vDSO that is used
when running cloudabi32 executables.

As arguments are automatically padded to 64-bit words by the vDSO in
userspace, we can copy the arguments directly into the array used by
truss(8) internally.

Reviewed by:	jhb
Differential Revision:	https://reviews.freebsd.org/D13516


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326911 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f