newsyslog: Fix issues after r326616
When building the command to execute for compression, newsyslog was modifying
the generic arguments array instead of its own copy.
Meaning on the second file to compress with the same arguments, the command line
was not the one expected.
Fix it by creating one copy of the arguments per execution and modifying that
copy.

While here, print the command line executed in verbose mode.

Reported by:	many


git-svn-id: svn+ssh://svn.freebsd.org/base/head@326930 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f