Avoid calling ${XCC} --version in every subdir
The caching in bsd.compiler.mk doesn't work when XCC is set on the command
line and XCFLAGS in the environment. Passing along the X_COMPILER_*
variables in crossenv ensures that bsd.compiler.mk doesn't check the
version of XCC in every subdirectory. For me this feels like a noticeable
speed up (especially when building on MacOS where process creation is more
expensive).