Add a cross-bootstrap-tools and cross-bootstrap-tools-install target
These targets will build all the bootstrap tools and FreeBSD tools and
then install them to BOOTSTRAP_TOOLS_DESTDIR. By default all the tool names
will be prefixed with `freebsd-` but this can be changed by setting
BOOTSTRAP_TOOLS_NAME_PREFIX.
The targets are useful when crossbuilding on Linux/MacOS to install all the
tools that are required in order to build e.g. a disk image into one
directory.