Initial changes to Makefile.inc1 to allow crossbuilding on Linux and Mac
When crossbuilding we can't use most of basic host commands since they have
slightly incompatible behaviour. Instead we need to build all tools during
the bootstrap-tools stage and then ensure that PATH only contains those
binariesinstead of the default host tools. For tools that are known to be
compatible with what the build systems expects (`cp`, `env`, etc.) we can
just add a symlink to $WORLDTMP instead of building them from the FreeBSD
sources.

In addition to this I also had to change all mtree invocations to include -W
when building on Linux/Mac since not all users and groups exist there.