Don't change directory owner to root when building with -DNO_ROOT
Summary:
Currently the mtree calls in Makefile.inc1 all change the directory owner
to match the spec file. However, we should not be doing this during
distributeworld if -DNO_ROOT is passed. Additionally, when creating the
WORLDTMP directory hierachy there is no need to change the owner to root so
we now always pass the -W flag when populating WORLDTMP.

This is also required for building FreeBSD on Linux/Mac since the required
groups/users will not exist there which is how I discovered this issue.

Reviewers: emaste, jhb, brooks, bdrewery, imp

Reviewed By: emaste, imp

Differential Revision: https://reviews.freebsd.org/D14185