Fix -Wunused-variable warning in bmake without USE_FILEMON
Summary:
This happens for example when building FreeBSD on MacOS since the check
in the Makefile will determine that filemon is not available. I am not
sure how to fix the check to be for the target system rather than the
host so for now I just added a TODO comment there.

Reviewers: sjg, imp, jhb

Reviewed By: imp

Differential Revision: https://reviews.freebsd.org/D14315