crossbuild: make mkioctls script work on Linux and MacOS
Summary:
Using find -s  will not work with the Linux or MacOS find command. We pipe
to sort instead since the only real requirement here is that the order
stays the same. While I am touching this file I also fixed a `==` construct
which is not supported by POSIX sh but appears to work on FreeBSD.

Reviewers: emaste, jhb, imp, bdrewery

Reviewed By: jhb, imp

Differential Revision: https://reviews.freebsd.org/D14246