Fix non-FreeBSD host lib32 build for TARGET=amd64
We need to pass an explicit target triple to clang otherwise it will
attempt to build with the host triple. This also has advantages when
building on a FreeBSD host: we now tell clang that we are targeting at least
FreeBSD 12.0 instead of an older version so it can enable newer features.