Fix calling convention mismatch in updateiov callback
This was causing a crash inside suword64() when calling sys_jail_set()
since user_jail_get() was passing a capability argument but the updateiov()
function was expecting an integer as the second argument. This then
resulted in a random capability being passed to suword64.

In my case this triggered an alignment trap, but the fswberr callback was
not being triggered. It turns out this is happening because pcb_onfault
is being cleared before being checked inside trap(). This will be fixed
in the next commit.